admins can now edit product info

+++ server/schemas/typeDefs.js
@@ -51,7 +51,7 @@ const typeDefs = gql`
   type Mutation {
     addUser(firstName: String!, lastName: String!, email: String!, password: String!): Auth
     addProduct(name: String!, description: String!, price:Float!, quantity:Int!, category:ID!): Product
-    editProduct(_id:ID!,name: String!, description: String!, price:Int!, quantity:Int!): Product
+    editProduct(_id:ID!,name: String!, description: String!, price:Float!, quantity:Int!): Product
     addOrder(products: [ID]!): Order
     updateUser(firstName: String, lastName: String, email: String, password: String): User
     updateProduct(_id: ID!, quantity: Int!): Product
